modification to ErrorObj to support ategory names longer than 20 characters (previously, longer names would not properly accept limits).
modification to MessageLoggerScribe to append .log (or .xml) to the names
of destination files if no extension is provided - this will be useful in
the context of the replace capability.
Also, careful checking to avoid the trap that comparison to the empty PSet
does not pick up on differences in untracked parameters.