Refactor the triggerCache logic into a separate function.  Makes it much more clear what is going on in the code.
As a part of the refactor, make sure that Branch->GetEntry is always
called.  The RelVal issues were caused because it was possible to construct
a scenario where this didn't happen (this scenario could only trigger for
very old RAW files, as that's a completely separate code path).  Beceause
we didnt actually invoke GetEntry, CMSSW was using undefined data.  This
refactoring should prevent this bug from being re-implemented in the future.

Only allow the trigger cache logic to function when we are in the event
tree.  This contributed to the bug described above.