Bug fix: handle indices from multiple BanchTypes
The previous code didn't check that the ProductHolderIndexHelper was for a particular BranchType when filling in the index value.
Since the Event info was filled last, any Run or Lumi indices obtained earlier would be reset to Unknown and therefore the framework
thought all Run or Lumi product requests were unregistered.