Remove CurrentProcessingContext from HLTrigger
Remove all uses of CurrentProcessingContext from HLTrigger
packages. This is replaced by using the new ModuleCallingContext.
There should be no change in behavior. The only interface change
in the HLTrigger code is that the Event is an added argument
to some functions. These functions are only used in a few
places and all uses are fixed. Also in the Framework, the
PlaceInPathContext class is added and used. Removed the unused
filling of TriggerResults for end paths. Some other minor
Framework cleanup and unit tests.