Separated internals of Schedule into GlobalSchedule and StreamSchedule
In order to accommodate multiple simultaneous Streams, the implementation of
Schedule has been broken into GlobalSchedule and StreamSchedule. The Schedule allows
for one StreamSchedule per Stream and at the moment only uses one GlobalSchedule.
The StreamSchedules and the GlobalSchedule share a ModuleRegistry in order to share module
instances.

The end job reports are a reduction of the summaries from all the StreamSchedules. The
GlobalSchedule does not participate since all of the information is Event based.

NOTE: This change is known not to pass all the unit tests.