Multithreaded TFileService
Modify TFileService so it will run in the multithreaded
Framework. Note that only "One" type modules that declare
TFileService as a shared resource will be able to use it
in multithreaded mode. TFileDirectory is now a thread local
member instead of being a base class of TFileService. This
change required some code that uses to the service to be
modified. Add Framework code that keeps track of the modules
active on the current thread and make this available to
the service through the context passed to the service.
Remove a couple unneeded includes in RecoParticleFlow
that were causing problems.