HistoryAppender no longer uses ProcessHistoryRegistry
To allow for thread-safe uses of the ProcessHistoryRegistry, the HistoryAppender no longer updates the ProcessHistoryRegistry to hold its appended history. Now it just returns a shared_ptr to the new or cached ProcessHistory. The shared_ptr is now owned by all Principals which are currently processing with that history.
To simplify the sharing, we only cache the last ProcessHistory and no longer hold a map to all previously seen histories. This will allow us to have one HistoryAppender per Stream when used by the SubProcess.