Enable NO_LIB_CHECKING on RFIODPMPlugin target
If SCRAM missing symbol check is enabled RFIODPMPlugin will fail
the check. The main library (Utilities/RFIOAdaptor) does not link
directly to castor, but has castor functions marked as extern in
./interface/RFIO.h. Thus castor needs to be loaded explicitly
before loading the library.

UtilitiesRFIOAdaptorPlugin passes as it has NO_LIB_CHECKING.
RFIOCastorPlugin passes as it links directly to castor.

This is minimal change. Otherwise we link to castor directly in
the main library as we do for Darwin.

Signed-off-by: David Abdurachmanov <David.Abdurachmanov@cern.ch>