Force creation of TStreamerInfos at TFile open time
The global container of TStreamerInfos is not thread-safe. Therefore the ROOT team advised to create all TStreamerInfos needed for a job up front. We already create all TStreamerInfos for schemas corresponding to the present class layouts. However, a TFile can contain old schemas and will wants to create their TStreamerInfos the first time the class is read from the file. For our purposes, that is too late and therefore we need to force ROOT to create them at file open time.