Remove getEngine() function from random service
In the random number generator service, remove the
getEngine function which takes no arguments. This function
is replaced by new getEngine functions that take either
a StreamID or a LuminosityBlockIndex as an argument. This
comes after a long migration of all clients to the new
interface. The new interace is designed to work with
the multithreaded Framework.

(Note that this also fixes a compilation error that resulted
after a similar pull request was automatically merged from
7_1_X)