Fix append bug in DQMEventMessage
The previous code obviously thought ‘std::string subFolderName = “Subfolder “ + idx;’ was appending ‘idx’ to the string. That is not what was happening. Instead, pointer math was happening. So if ‘idx’ was 1 the resulting string was “ubfolder “. If ‘idx’ was 2 we got ‘bfolder ‘. If ‘idx’ was larger than the string length we would get random garbage.
This problem was found by clang.