AssociationMap must be initialized and can use Ptr and RefToBase for lookup
Changed AssocationMap so that it is possible to lookup items use a
Ptr or a RefToBase. This avoids requiring users to create a edm::Ref
just for lookup.
Added a constructor to AssociationMap to allow one to associate which
collection(s) the container will use for association. This constructor
must be used since it is no longer possible to have those collections
be set on the first insert to the container.