No longer create a edm::RefProd from a edm::Ref
The design which allowed a edm::RefProd to be constructed from an
edm::Ref caused problems for future changes (edm::View and slimming).
Given this was equivalent to taking an arbitrary pointer and then
obtaining which std::vector<> the data is stored in, we determined
this was not an absolute requirement.