No longer create a edm::RefToBaseProd from a edm::Ref or RefToBase
The design which allowed a RefToBaseProd to be constructed from an
edm::Ref or RefToBase caused problems for future changes (edm::View
and slimming). Given this was equivalent to taking an arbitrary pointer
and then obtaining which std::vector<> the data is stored in, we
determined this was not an absolute requirement.