Add new storage statistics service for Xrootd.
This service keeps per-site statistics from the Xrootd layer.
Previously, the IO statistics were kept by the StatisticsSenderService
on a per-file level.  These are less-useful now that Xrootd reads
from multiple servers (and possibly multiple sites) at a time!

Currently, this has to be enabled manually and only outputs
to the FJR.  Long-term, we plan on adding additional statistics
and additional output mechanisms (correcting the now-broken
StatisticsSenderService reports we send to the Dashboard).

NOTE: there is some trickery here due to the fact that the service
registry is only available on CMSSW-managed threads.  In order to
make the service available from the callback threads, the service
pointer is cached by the request manager object.