Check immediate return status from XrdClFile::Open call.
This catches a case where the return status from XrdClFile::Open
is not checked.  Previously, oversight this was OK -- no failure
could happen immediately.  However, in bdbec57a we introduced a
query to the redirector prior to the first file open.  If the query
failed with a fatal error, then the open may fail immediately.

As the open status is not checked, we proceed to waiting for the
callback - which hangs indefinitely.