Make Timing service report properly when using threads
CPU time measurement is for the application as a whole and therefore cannot be determined on an event by event basis. Therefore the event level CPU measurement has been dropped.
The average event time is now properly calculated so it is the average time it takes an event to processes on a stream.
We now calculate event throughput which measures how fast the entire job is processing events on average.