Fix sporadically failing unit tests
Fix sporadically failing unit tests in
DataFormats/FWLite. This only affects
unit tests.

Previously, the test forked a process to
make input files and then was supposed to
wait for it to finish, but sometimes it
did not wait for reasons I do not understand.
I simplified this by splitting it into two
tests and used the scram PRE_TEST flag to
sequence them.

The DataFormats/FWLite tests used the same
filenames as the FWCore/FWLite unit tests
in the same directory. I renamed the files
and modified the test to run in the tmp
subdirectory to avoid potential clashes if
the tests were run at the same time.