Replace use of pstack with direct invocation of gdb.
This switches from the system pstack script to using gdb directly.
Ideally, this will be more usable on non-Linux platforms.

NOTE that this will pull the gdb from the environment (i.e., probably
CMSSW's gdb instead of the system one).  CMSSW's gdb can parse the
debug symbols we ship, meaning that the backtrace will include
niceties like filenames, line numbers, and decode arguments.

The downside of this is that CMSSW's gdb requires 10 minutes
and about 1GB RAM to generate the traceback.