Send storage statistics to HTCondor.
This commit improves the CondorUpdateService to send storage statistics
too.

This required a refactor / cleanup of the StorageAccount class to expose
various statistics to the outside world in a thread-safe manner (previously,
all accesses were not-thread-safe).  As a part of this, we improved the
internals to prevent floating point underflow.

This also fixes Xrootd reporting to the StatisticsSenderService.