Fix thread-safety issue with StreamerInputSource
The previous way StreamerInputSource handled reading back RefCores
lead to the probability that when dereferenced the RefCore would read
from the wrong EventPrincipal when the job was using multiple streams
becuase it was reusing the same EventPrincipalHolder for each event.
The code has now been changed to make a new EventPrincipalHolder for
each event.