Removed Transient from ProductProvenance
Transient was primarily used to do a cache of edm::Parentage even though the information is stored in ParentageRegistry. The Parentage object is only requested on output and only once per branch so the cache doesn't look to be all that useful. The cache is sometimes updated lazily which causes thread-safety issues. The easiest solution is just to remove the use of the Transient all together.
(cherry picked from commit 877e7be91c4e72608d28d3d69e1c171829cb9bc2)