Process.extend now applies ProcessModifiers last
During an extend call we need to apply all ProcesModifiers at the end of the call. That way if the modifier references an item from the Process we will be sure it has already been attached to the Process.
(cherry picked from commit e4bfdc7fb814a1684ac95ad5701a178834522a42)