Allow FileSystem queries to occur from callback threads.
To query a filesystem property, a free callback thread is needed
with the default synchronous versions of XrdCl::FileSystem methods.

Thus, if the synchronous method is called from a *callback thread*,
then there must be an idle thread to handle the response; if there
are N threads in the pool and N simultaneous queries, then the
filesystem query would cause a deadlock.

Worse yet, the timeout mechanism for XrdCl relies on an idle callback
thread to function.

As making the synchronous query asynchronous in its currrent use case
is quite hard, I only partially solved the problem: wrote a new
response handler that can timeout without an idle thread.