Refactor memory management protocol for QueryAttrHandler.
All "real" state is kept in a sub-object; the QueryAttrHandler keeps
a weak_ptr to the "real" state.  While the main thread is waiting,
it keeps a reference to the shared_ptr for the state, keeping the
weak_ptr alive.

If the timeout occurs, the shared_ptr is dropped and the weak_ptr
cannot be locked.

The QueryAttrHandler is always deleted by the callback if it is
successfully registered.