Correct lifetime of outstanding open-file request.
Previously, the presence of a pointer in m_file indicated an open-
file request was outstanding.  However, after the file was opened,
the unique_ptr was moved to the Source object (which does a synchronous
network query in its constructor).  Hence, it appeared there was
no outstanding file-open and a second one could start.

Now, we keep a separate boolean flag that is only set to false when
the HandleResponse method exits.
(cherry picked from commit 2ef8f99d56099c444e1ef58b0007881b59c42865)