Only construct tbb::task_scheduler_init once
With the use of TBB version 44_20160316oss, if we create a tbb::task_scheduler_init, then destroy that instance and create another one, we get a segmentation fault because TBB's global state is not properly reset. It is unclear if this is an allowed use case or not. For now (and possible forever) we will only make one per job.